staging/app/requirements.txt
Flask
mysql-connector
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Cors
requests
pika
flask-mysql
PyMySQL
cryptography
mysql-connector-python